Standardise duplicated Slim* structs and move them into API models, to allow consistent implementation of EditRoute on them where required
