+++ thoth-api/src/contribution/model.rs
@@ -9,6 +9,7 @@ use uuid::Uuid;
 use crate::schema::contribution;
 #[cfg(feature = "backend")]
 use crate::schema::contribution_history;
+use crate::work::model::SlimWork;
 
 #[cfg_attr(feature = "backend", derive(DbEnum, juniper::GraphQLEnum))]
 #[cfg_attr(feature = "backend", DieselType = "Contribution_type")]
@@ -69,6 +70,12 @@ pub struct Contribution {
     pub first_name: Option<String>,
     pub last_name: String,
     pub full_name: String,
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SlimContribution {
+    pub work: SlimWork,
 }
 
 #[cfg_attr(
