+++ thoth-api/src/imprint/model.rs
@@ -8,6 +8,7 @@ use uuid::Uuid;
 
 use crate::graphql::utils::Direction;
 use crate::publisher::model::Publisher;
+use crate::publisher::model::SlimPublisher;
 #[cfg(feature = "backend")]
 use crate::schema::imprint;
 #[cfg(feature = "backend")]
@@ -50,6 +51,12 @@ pub struct ImprintExtended {
     pub imprint_url: Option<String>,
     pub updated_at: DateTime<Utc>,
     pub publisher: Publisher,
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SlimImprint {
+    pub publisher: SlimPublisher,
 }
 
 #[cfg_attr(
