+++ thoth-api/src/series/model.rs
@@ -74,6 +74,13 @@ pub struct SeriesExtended {
     pub imprint: Imprint,
 }
 
+#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SlimSeries {
+    pub series_id: Uuid,
+    pub series_name: String,
+}
+
 #[cfg_attr(
     feature = "backend",
     derive(juniper::GraphQLInputObject, Insertable),
