+++ thoth-app/src/models/contributor/contributor_activity_query.rs
@@ -1,5 +1,6 @@
 use serde::Deserialize;
 use serde::Serialize;
+use thoth_api::contribution::model::SlimContribution;
 use uuid::Uuid;
 
 use crate::graphql_query_builder;
@@ -13,6 +14,7 @@ pub const CONTRIBUTOR_ACTIVITY_QUERY: &str = "
                     title
                     imprint {
                         publisher {
+                            publisherId
                             publisherName
                         }
                     }
@@ -48,30 +50,4 @@ pub struct ContributorActivityResponseData {
 #[serde(rename_all = "camelCase")]
 pub struct ContributorActivity {
     pub contributions: Option<Vec<SlimContribution>>,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct SlimContribution {
-    pub work: SlimWork,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct SlimWork {
-    pub work_id: Uuid,
-    pub title: String,
-    pub imprint: SlimImprint,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct SlimImprint {
-    pub publisher: SlimPublisher,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct SlimPublisher {
-    pub publisher_name: String,
 }
