+++ thoth-app/src/models/publication/delete_publication_mutation.rs
@@ -14,9 +14,12 @@ const DELETE_PUBLICATION_MUTATION: &str = "
             publicationType
             workId
             work {
+                workId
+                title
                 imprint {
                     publisher {
                         publisherId
+                        publisherName
                     }
                 }
             }
