+++ thoth-app/src/models/publication/publication_query.rs
@@ -20,9 +20,12 @@ pub const PUBLICATION_QUERY: &str = "
                 updatedAt
             }
             work {
+                workId
+                title
                 imprint {
                     publisher {
                         publisherId
+                        publisherName
                     }
                 }
             }
